requirements.txt
asgiref==3.3.0
certifi==2020.11.8
chardet==3.0.4
dj-database-url==0.5.0
Django==3.1.3
django-embed-video==1.3.3
django-heroku==0.3.1
gunicorn==20.0.4
idna==2.10
psycopg2==2.8.6
pytz==2020.4
requests==2.25.0
sqlparse==0.4.1
urllib3==1.26.2
whitenoise==5.2.0
